Extract screenshot into new component, move images
