+++ src/data/projects.json
@@ -10,7 +10,7 @@
       "The video below shows a demonstration of the final application."
     ],
     "videoSource": "riQTxIDijEs",
-    "imageSource": "project-images/mass-spring-particle-system.jpg"
+    "imageSource": "mass-spring-particle-system.jpg"
   },
   {
     "id": 2,
@@ -19,7 +19,7 @@
     "blurb": "Coming soon",
     "details": [],
     "videoSource": "",
-    "imageSource": "project-images/mobile-game-coming-soon.jpg"
+    "imageSource": "mobile-game-coming-soon.jpg"
   },
   {
     "id": 3,
